requirements.txt
streamlit>=1.28.1
pymongo>=4.6.0
sentence-transformers>=2.2.2
openai>=1.6.1
PyPDF2>=3.0.1
python-docx>=0.8.11
numpy>=1.24.3
pandas>=2.0.3
python-dotenv>=1.0.0
